requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
motor>=3.1.1
pymongo>=4.3.3
redis>=4.5.4
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
anthropic>=0.5.0
email-validator>=2.0.0